pyproject.toml
[project]
name = "telegram-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "motor>=3.7.1",
    "pymongo>=4.16.0",
    "telethon>=1.42.0",
    "python-dotenv>=1.0.1"
]
